update scan classes
idx coords are ints now
load() now loads point scans

+++ src/trion/expt/scans.py
@@ -256,7 +256,7 @@ class ContinuousScan(BaseScan):
         self.afm_data = xr.Dataset()
         tracked_channels = ['idx', 'x', 'y', 'z', 'amp', 'phase']
         for i, c in enumerate(tracked_channels[1:]):
-            da = xr.DataArray(data=afm_tracking[:, i+1], dims='idx', coords={'idx': afm_tracking[:, 0]})
+            da = xr.DataArray(data=afm_tracking[:, i+1], dims='idx', coords={'idx': afm_tracking[:, 0].astype('int')})
             if self.t is not None:
                 da = da.expand_dims(dim={'t': np.array(self.t)})
             self.afm_data[c] = da
